// packages/sshoc/lib/types.ts
export interface paths {
// ...
    "/api/item-search": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description Search among items. */
        get: operations["searchItems"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        WorkflowCore: {
            label: string;
// ...
        };
// ...
        PaginatedSearchItems: {
            /** Format: int64 */
            hits?: number;
            /** Format: int32 */
            count?: number;
            /** Format: int32 */
            page?: number;
            /** Format: int32 */
            perpage?: number;
            /** Format: int32 */
            pages?: number;
            q?: string;
            order?: ("score" | "label" | "modified-on")[];
            items?: components["schemas"]["SearchItem"][];
            categories?: {
                [key: string]: components["schemas"]["LabeledCheckedCount"];
            };
            facets?: {
                [key: string]: {
                    [key: string]: components["schemas"]["CheckedCount"];
                };
            };
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    getWorkflow: {
        parameters: {
            query?: {
                draft?: boolean;
                approved?: boolean;
                redirect?: boolean;
            };
            header?: never;
            path: {
                persistentId: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["WorkflowDto"];
                };
            };
        };
    };
// ...
    getActor: {
        parameters: {
            query?: {
                items?: boolean;
            };
            header?: never;
            path: {
                id: number;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ActorDto"];
                };
            };
        };
    };
// ...
    searchItems: {
        parameters: {
            query?: {
                q?: string;
                categories?: ("tool-or-service" | "training-material" | "publication" | "dataset" | "workflow" | "step")[];
                order?: ("score" | "label" | "modified-on")[];
                page?: number;
                perpage?: number;
                advanced?: boolean;
                includeSteps?: boolean;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["PaginatedSearchItems"];
                };
            };
        };
    };
// ...
}